feature: Cosmetics and ownership state (#11)
Adds the ability to query cosmetics and the ownership state of them for players.

Note that the inability to query cosmetics without a player is intentional as not every player has access to every cosmetic. Therefore, cosmetics that players do not have access to should not be counted as cosmetics that exist for them!