chore: Deprecate gems for removal
Looking up the amount of gems a player has will now always return 0 until such a time that we remove it. We've had a fair few complaints that it allows people to "spy" on when/how many purchases are being made for players and that's not something we want.

Also fixes a typo in the docs for chroma packs.